fix more x64 compiler warnings

+++ baseutils/str_util.c
@@ -31,8 +31,8 @@ void strcpy_s(char *dst, size_t dstLen, const char *src)
 char * str_cat_s(char * dst, size_t dst_cch_size, const char * src)
 {
     size_t len = str_len(dst);
-    int count = dst_cch_size - len;
-    int ret = _snprintf(dst + len, count, "%s", src);
+    size_t count = dst_cch_size - len;
+    size_t ret = _snprintf(dst + len, count, "%s", src);
     return (ret<count ) ? dst : NULL;
 }
